movie-app/movie-app-backend/requirements.txt
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.1.3
python-dotenv
